Move tests for client.client.verify_show_unpublished() into their own file
